pipelines/dynamo-db-fanout-sns/requirements.txt
pytest
aws-cdk.core==1.44.0
aws-cdk.aws-dynamodb==1.44.0
aws-cdk.aws-iam==1.44.0
aws-cdk.aws-lambda==1.44.0
aws-cdk.aws-lambda-destinations==1.44.0
aws-cdk.aws-lambda-event-sources==1.44.0
aws-cdk.aws-sns==1.44.0
aws-cdk.aws-sns-subscriptions==1.44.0
aws-cdk.aws-sqs==1.44.0
